+++ src/view/console.rs
@@ -5,7 +5,8 @@ use colored::*;
 use itertools::Itertools;
 
 use solver::{
-    task::{Goal, GoalKind, Solution, TermProps},
+    engine::{Solution, TermProps},
+    task::{Goal, GoalKind},
     term::SharedTerm,
 };